src/lambda_codebase/account_processing/requirements.txt
pyyaml
wrapt==1.12
aws-xray-sdk